fix: Rename RPC methods to avoid naming conflicts with message types
Changes:
- MediaHandlerService.RegisterParticipant -> Register
- MeetingControllerService.AssignMeeting -> Assign

Reason: Protocol Buffers does not allow RPC methods to have the same
name as message types. This was causing protoc compilation errors:
"RegisterParticipant is not a message type"
"AssignMeeting is not a message type"

The message types are correct and should not be renamed. RPC methods
now use verb naming (Register, Assign) which is more idiomatic.